feat(apiutils): GenericApiErrorResponseHandler created
