+++ apps/ui/src/api/queries/useGetUser.ts
@@ -20,6 +20,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
 
+import { GenericApiErrorResponseHandler } from '@/api/apiUtils';
 import { fetch } from '@/global/FetchClient';
 import { ServerError } from '@/global/types';
 import type { UserResponse } from '@pcgl-daco/validation';
@@ -32,21 +33,7 @@ const useGetUser = () => {
 		queryFn: async () => {
 			const response = await fetch(`/auth/user`);
 
-			if (!response.ok) {
-				const error = {
-					message: translate('errors.generic.title'),
-					errors: translate('errors.generic.message'),
-				};
-
-				switch (response.status) {
-					case 400:
-						error.message = translate('errors.http.400.title');
-						error.errors = translate('errors.http.400.message');
-						break;
-				}
-
-				throw error;
-			}
+			GenericApiErrorResponseHandler({ response, translate });
 
 			return await response.json();
 		},
